jester_backend/requirements.txt
Django==1.7.2
django-cors-headers==1.0.0
django-enumfield==1.2
django-ipware==0.1.0
enum==0.4.4
matplotlib==1.4.2
mock==1.0.1
MySQL-python==1.2.5
nose==1.3.4
numpy==1.9.1
pyparsing==2.0.3
python-dateutil==2.4.0
pytz==2014.10
scikit-learn==0.15.2
scipy==0.14.1
six==1.9.0
sqlparse==0.1.14
